requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
exceptiongroup==1.2.1
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.5.0
gradio-client==0.16.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.1
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
packaging==24.0
pydantic==2.5.3
pydantic-core==2.14.6
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.2
sniffio==1.3.1
tqdm==4.66.4
typing-extensions==4.10.0
urllib3==2.2.1
websockets==11.0.3
yarl==1.9.4
